requirements.txt
# ---------------- Core Framework ----------------
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# ---------------- Database & ORM ----------------
sqlalchemy>=2.0.36,<2.1.0
# asyncpg==0.29.0  # (Optional Postgres driver) commented out to avoid C build on Windows Python 3.13
alembic==1.12.1
sqlite-utils==3.35.1

# ---------------- LangGraph & AI ----------------
langgraph==0.1.9
langchain-core==0.2.29
openai==1.3.7
qdrant-client==1.9.0
tenacity==8.2.3

# ---------------- Web Scraping -------------------
# aiohttp==3.9.1      # optional (web scraping async HTTP) - disabled for minimal runtime
# selectolax==0.3.17  # optional (HTML parsing) - disabled for minimal runtime
beautifulsoup4==4.12.2

# ------------- Scheduling / Jobs -----------------
apscheduler==3.10.4

# --------------- Auth & Security -----------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# --------------- Rate Limiting -------------------
slowapi==0.1.9

# ------------- Config & Environment --------------
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
python-dotenv==1.0.0
pyyaml==6.0.1

# --------------- Data Processing -----------------
# ---------------- Optional Data Tools (commented out to avoid build issues on Python 3.13 without C toolchain) ---------
# pandas==2.1.4  # Uncomment if you need data loading scripts; may require installing build tools or using a wheel-compatible Python version
# numpy pinned version removed due to Python 3.13 build issues; uncomment or specify a newer wheel-supported version if needed
# numpy==1.24.4
python-dateutil==2.8.2
pytz==2023.3

# ------------------- Testing ---------------------
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# ------------- Linting / Formatting --------------
ruff==0.1.6

# ------------------- Logging ---------------------
structlog==23.2.0

# --------------- HTTP Client misc ----------------
requests==2.31.0
